Move all component logic out of `:app`
