+++ app/build.gradle
@@ -28,17 +28,10 @@ dependencies {
     implementation project(":core:theme")
     implementation libs.androidx.activity
     implementation libs.androidx.activityCompose
-    implementation libs.androidx.compose.animation
     implementation libs.androidx.compose.runtime
-    implementation libs.androidx.compose.ui
     implementation libs.androidx.fragment
-    implementation libs.androidx.lifeCycleViewModel
-    implementation libs.androidx.lifeCycleViewModelSavedState
-    implementation libs.androidx.navigation.compose
-    implementation libs.androidx.navigationCommon
-    implementation libs.androidx.navigationRuntime
-    implementation libs.androidx.viewmodelCompose
-    implementation libs.hilt.navigationCompose
+    implementation libs.androidx.viewmodel
+    implementation libs.androidx.viewmodelSavedState
     implementation libs.kotlinx.serializationCore
 
     debugImplementation libs.androidx.compose.uiTooling
