+++ core/components-ui/src/main/kotlin/uk/co/jordanterry/components/ui/ComponentScreen.kt
@@ -3,11 +3,16 @@ package uk.co.jordanterry.components.ui
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
+import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
+import androidx.navigation.compose.rememberNavController
 import uk.co.jordanterry.components.ui.system.SingleColumnListScreen
 import uk.co.jordanterry.core.components.Layout
+import uk.co.jordanterry.core.components.Screen
 
 @Composable
-public fun ComponentScreen(
+private fun ComponentScreen(
     componentViewModel: ComponentViewModel,
 ) {
     val state by componentViewModel.state.collectAsState()
@@ -15,5 +20,17 @@ public fun ComponentScreen(
         is Layout.SingleColumnList -> SingleColumnListScreen(
             layout
         )
+    }
+}
+
+@Composable
+public fun ComponentScreen() {
+    val navController = rememberNavController()
+    NavHost(navController = navController, startDestination = Screen("home")) {
+        composable<Screen> { _ ->
+            val componentViewModel: ComponentViewModel =
+                hiltViewModel()
+            ComponentScreen(componentViewModel)
+        }
     }
 }
