+++ core/components-ui/src/main/kotlin/uk/co/jordanterry/components/ui/system/Components.kt
@@ -56,7 +56,9 @@ internal fun ComponentButton(
     modifier: Modifier = Modifier,
 ) {
     Button(
-        onClick = {},
+        onClick = {
+            button.action
+        },
         modifier = modifier,
     ) { Text(button.text) }
 }
