+++ gradle/libs.versions.toml
@@ -40,10 +40,10 @@ androidx-compose-uiUnit = { group = "androidx.compose.ui", name = "ui-unit" }
 androidx-composeBom = { group = "androidx.compose", name = "compose-bom", version.ref = "composeBom" }
 androidx-core = { group = "androidx.core", name = "core-ktx", version.ref = "androidx-core" }
 androidx-fragment = { group = "androidx.fragment", name = "fragment", version.ref = "androidx-fragment" }
-androidx-lifeCycleViewModel = { group = "androidx.lifecycle", name = "lifecycle-viewmodel", version.ref = "androidx-lifecycle" }
-androidx-lifeCycleViewModelSavedState = { group = "androidx.lifecycle", name = "lifecycle-viewmodel-savedstate", version.ref = "androidx-lifecycle" }
+androidx-viewmodel = { group = "androidx.lifecycle", name = "lifecycle-viewmodel", version.ref = "androidx-lifecycle" }
+androidx-viewmodelSavedState = { group = "androidx.lifecycle", name = "lifecycle-viewmodel-savedstate", version.ref = "androidx-lifecycle" }
 androidx-lifecycleRuntime = { group = "androidx.lifecycle", name = "lifecycle-runtime-ktx", version.ref = "androidx-lifecycle" }
-androidx-navigation-compose = { group = "androidx.navigation", name = "navigation-compose", version.ref = "androidx-navigationCompose" }
+androidx-navigationCompose = { group = "androidx.navigation", name = "navigation-compose", version.ref = "androidx-navigationCompose" }
 androidx-navigationCommon = { group = "androidx.navigation", name = "navigation-common", version.ref = "androidx-navigationCompose" }
 androidx-navigationRuntime = { group = "androidx.navigation", name = "navigation-runtime", version.ref = "androidx-navigationCompose" }
 androidx-viewmodelCompose = { group = "androidx.lifecycle", name = "lifecycle-viewmodel-compose", version.ref = "androidx-lifecycle" }
@@ -52,7 +52,7 @@ dagger-hiltAndroid = { group = "com.google.dagger", name = "hilt-android", versi
 dagger-hiltAndroidCompiler = { group = "com.google.dagger", name = "hilt-android-compiler", version.ref = "dagger" }
 dagger-hiltCompiler = { group = "com.google.dagger", name = "hilt-compiler", version.ref = "dagger" }
 dagger-hiltCore = { group = "com.google.dagger", name = "hilt-core", version.ref = "dagger" }
-hilt-navigationCompose = { group = "androidx.hilt", name = "hilt-navigation-compose", version.ref = "hiltNavigationComposeVersion" }
+dagger-hilt-navigationCompose = { group = "androidx.hilt", name = "hilt-navigation-compose", version.ref = "hiltNavigationComposeVersion" }
 javax-inject = { group = "javax.inject", name = "javax.inject", version = "javax-inject" }
 kotlinx-coroutinesCore = { group = "org.jetbrains.kotlinx", name = "kotlinx-coroutines-core", version.ref = "kotlinx-coroutines" }
 kotlinx-serializationCore = { group = "org.jetbrains.kotlinx", name = "kotlinx-serialization-core", version.ref = "kotlinx-serialization" }
